fix: adding attacks only to cells with enemy pieces
Fix the fact that add_attacks_in_direction_to and thus all the function which
rely on it marking cells as attacked only when there's a piece at that
cell and this piece is of enemy color.

The function is expected to add all the cells which the given piece can
attack, even if the cells are empty.